+++ src/main/java/model/MatchMapper.java
@@ -1,0 +1,16 @@
+package oit.is.z0793.kaizi.janken.model;
+
+import java.util.ArrayList;
+
+import org.apache.ibatis.annotations.Insert;
+import org.apache.ibatis.annotations.Mapper;
+import org.apache.ibatis.annotations.Options;
+import org.apache.ibatis.annotations.Select;
+
+@Mapper
+public interface MatchMapper {
+
+  @Select("SELECT * from matches")
+  ArrayList<Match> selectAllMatch();
+
+}
